refactor: reuse artifacts query

+++ posthog/schema.py
@@ -8167,7 +8167,7 @@ class CodebaseTreeQuery(BaseModel):
         default=None, description="Modifiers used when performing the query"
     )
     response: Optional[CodebaseTreeQueryResponse] = None
-    userId: float
+    userId: int
 
 
 class Response3(BaseModel):
@@ -8237,7 +8237,7 @@ class EditorSemanticSearchQuery(BaseModel):
         default=None, description="Modifiers used when performing the query"
     )
     response: Optional[EditorSemanticSearchQueryResponse] = None
-    userId: float
+    userId: int
 
 
 class EventTaxonomyQuery(BaseModel):
@@ -8593,7 +8593,7 @@ class SyncedArtifactsQuery(BaseModel):
         default=None, description="Modifiers used when performing the query"
     )
     response: Optional[SyncedArtifactsQueryResponse] = None
-    userId: float
+    userId: int
 
 
 class TeamTaxonomyQuery(BaseModel):
